Smart-KIOS/ai/requirements.txt
Flask
mysql-connector-python
transformers
torch
Flask-Cors
gtts
torchaudio